First commit: Import 1.0 ara-web skeleton
In ARA 1.0, the different components that makes ARA are being split
into their own projects, modules and repositories.

This will, amongst other things, allow for the different components to
be installed in a modular fashion only where they make sense.

ara-web will be the repository and project which will contain the
frontend web-facing interface.

This is a first commit/import from what is roughly a skeleton and we'll
iterate on that.

Change-Id: Id035c2eb9d0fe55b3cada6daf04d94fc76ba95f0