Leverage refactored API for displaying playbook details
Numbers for the summary are now provided, we no longer need to count them.

We no longer need dedicated task actions (for now), remove them.

In the playbook details, we now iterate through each play to print a table that reconciles
tasks, results and hosts involved in that play.

The design of the UI is not final and is mostly a proof of concept for
the visualizing the refactored API.

Depends-On: https://review.openstack.org/#/c/641136/
Change-Id: I27180fab04e1192516be6f12fee53c299f434fee